LPS-176298 Extract to a new class
