+++ modules/apps/info/info-api/src/main/java/com/liferay/info/field/type/MultiselectInfoFieldType.java
@@ -14,10 +14,7 @@
 
 package com.liferay.info.field.type;
 
-import com.liferay.info.localized.InfoLocalizedValue;
-
 import java.util.Collection;
-import java.util.Locale;
 
 /**
  * @author Víctor Galán
@@ -27,50 +24,13 @@ public class MultiselectInfoFieldType implements InfoFieldType {
 	public static final MultiselectInfoFieldType INSTANCE =
 		new MultiselectInfoFieldType();
 
-	public static final Attribute<MultiselectInfoFieldType, Collection<Option>>
-		OPTIONS = new Attribute<>();
+	public static final Attribute
+		<MultiselectInfoFieldType, Collection<OptionInfoFieldType>> OPTIONS =
+			new Attribute<>();
 
 	@Override
 	public String getName() {
 		return "multiselect";
-	}
-
-	public static class Option {
-
-		public Option(
-			boolean active, InfoLocalizedValue<String> labelInfoLocalizedValue,
-			String value) {
-
-			_active = active;
-			_labelInfoLocalizedValue = labelInfoLocalizedValue;
-			_value = value;
-		}
-
-		public Option(
-			InfoLocalizedValue<String> labelInfoLocalizedValue, String value) {
-
-			_labelInfoLocalizedValue = labelInfoLocalizedValue;
-			_value = value;
-
-			_active = false;
-		}
-
-		public String getLabel(Locale locale) {
-			return _labelInfoLocalizedValue.getValue(locale);
-		}
-
-		public String getValue() {
-			return _value;
-		}
-
-		public boolean isActive() {
-			return _active;
-		}
-
-		private final boolean _active;
-		private final InfoLocalizedValue<String> _labelInfoLocalizedValue;
-		private final String _value;
-
 	}
 
 	private MultiselectInfoFieldType() {
