+++ modules/apps/info/info-api/src/main/java/com/liferay/info/field/type/SelectInfoFieldType.java
@@ -14,10 +14,7 @@
 
 package com.liferay.info.field.type;
 
-import com.liferay.info.localized.InfoLocalizedValue;
-
 import java.util.Collection;
-import java.util.Locale;
 
 /**
  * @author Alicia Garcia
@@ -30,50 +27,13 @@ public class SelectInfoFieldType implements InfoFieldType {
 	public static final SelectInfoFieldType INSTANCE =
 		new SelectInfoFieldType();
 
-	public static final Attribute<SelectInfoFieldType, Collection<Option>>
-		OPTIONS = new Attribute<>();
+	public static final Attribute
+		<SelectInfoFieldType, Collection<OptionInfoFieldType>> OPTIONS =
+			new Attribute<>();
 
 	@Override
 	public String getName() {
 		return "select";
-	}
-
-	public static class Option {
-
-		public Option(
-			boolean active, InfoLocalizedValue<String> labelInfoLocalizedValue,
-			String value) {
-
-			_active = active;
-			_labelInfoLocalizedValue = labelInfoLocalizedValue;
-			_value = value;
-		}
-
-		public Option(
-			InfoLocalizedValue<String> labelInfoLocalizedValue, String value) {
-
-			_labelInfoLocalizedValue = labelInfoLocalizedValue;
-			_value = value;
-
-			_active = false;
-		}
-
-		public String getLabel(Locale locale) {
-			return _labelInfoLocalizedValue.getValue(locale);
-		}
-
-		public String getValue() {
-			return _value;
-		}
-
-		public boolean isActive() {
-			return _active;
-		}
-
-		private final boolean _active;
-		private final InfoLocalizedValue<String> _labelInfoLocalizedValue;
-		private final String _value;
-
 	}
 
 	private SelectInfoFieldType() {
